Turned off capture of POSIX signals.
Since the catch2 environment changed from headers only to a static library,
the POSIX signals cannot be turned ON or OFF within a specific test. Instead,
we have to compile the library that way. It is neat that catch2 can capture
the signal, but that prevents gdb from doing so, which can be much more
helpful since gdb gives us the full state at the point the signal happened.